+++ application/modules/common/controllers/UserDefaultController.php
@@ -501,8 +501,6 @@ class UserDefaultController extends Zend_Controller_Action
                 $user->setModificationDate();
                 $screenName = $user->getScreenName();
 
-                $user->setApiPassword(password_hash(Ccsd_Tools::generatePw(), PASSWORD_DEFAULT));
-
                 if ($user->save()) {
                     $success = Zend_Registry::get('Zend_Translate')->translate("L'utilisateur <strong>%%RECIPIENT_SCREEN_NAME%%</strong> a bien été ajouté à Episciences");
                     $success = str_replace('%%RECIPIENT_SCREEN_NAME%%', $screenName, $success);
@@ -565,9 +563,9 @@ class UserDefaultController extends Zend_Controller_Action
 
                 $user->setValid(0);
                 $user->setIs_valid(); // Episciences validation
-                $user->setApiPassword(password_hash(Ccsd_Tools::generatePw(), PASSWORD_DEFAULT));
 
                 $lastInsertId = $user->save();
+
                 try {
                     $user->setUid($lastInsertId);
                 } catch (Exception $e) {
@@ -737,7 +735,6 @@ class UserDefaultController extends Zend_Controller_Action
                 $user = new Episciences_User($updatedUserValues);
 
 
-
                 $subform = $form->getSubForm('ccsd');
 
                 if ($subform->PHOTO->isUploaded()) {
@@ -746,7 +743,12 @@ class UserDefaultController extends Zend_Controller_Action
 
                     try {
                         $user->savePhoto($photoFileName);
-                        Episciences_Auth::incrementPhotoVersion();
+                        $user->savePhotoWithUuid($photoFileName);
+
+                        if ($user->getUid() === Episciences_Auth::getUid()) {
+                            Episciences_Auth::incrementPhotoVersion();
+                        }
+
                     } catch (Exception $e) {
                         $this->_helper->FlashMessenger->setNamespace('danger')->addMessage($e->getMessage());
                     }
@@ -1443,6 +1445,7 @@ class UserDefaultController extends Zend_Controller_Action
         if ($this->getRequest()->isXmlHttpRequest() && isset($params['uid'])) {
             if (Episciences_Auth::getUid() == $params['uid'] || Episciences_Auth::isSecretary()) {
                 $user = new Ccsd_User_Models_User(['uid' => $params['uid']]);
+                $user->setUuid(Episciences_UserManager::getUuidFromUid($params['uid']));
                 $user->deletePhoto();
                 Episciences_Auth::incrementPhotoVersion();
                 if (Episciences_Auth::getUid() == $params['uid']) {
@@ -1762,14 +1765,14 @@ class UserDefaultController extends Zend_Controller_Action
         unset($localUserData['ROLES'], $localUserData['affiliations'], $localUserData['web_sites'], $localUserData['biography'], $localUserData['social_medias']); // to fix PHP Notice: Array to string conversion
         $res = $user->findWithCAS($user->getUid());
 
-        if($res === null){
+        if ($res === null) {
             trigger_error("This account could not be found.", E_USER_ERROR);
         }
 
         $casUserData = $user->toArray();
-        unset($casUserData['ROLES'], $casUserData['affiliations'], $localUserData['biography'] , $casUserData['web_sites'], $casUserData['social_medias']);
+        unset($casUserData['ROLES'], $casUserData['affiliations'], $localUserData['biography'], $casUserData['web_sites'], $casUserData['social_medias']);
 
-        if(!empty(array_diff($localUserData, $casUserData))){
+        if (!empty(array_diff($localUserData, $casUserData))) {
             $data = array_merge($localUserData, $casUserData);
             $user = new Episciences_User($data);
             $user->save(false, false);